refactor(mespapiers): Move render out of PapersList
to isolate queries logic
